frontend/requirements.txt
gradio
requests
Pillow
numpy